Add files via upload
This is a QuickSight dashboard of one of my final projects from my Data Exploration and Visualization class, in which I obtained and processed publicly available data from data.gov regarding electric vehicles registered in the state of Washington in the United States.  I compiled all of the sheets in a .RAR directory for ease of access.  To extract the files in the directory, install either WinRAR or WinZip, right-click the directory, and click extract.