Add files via upload
I worked on a relational and SQL database design project for my database systems class.  The idea was to create a database that could compile pricing data from various health insurance providers.  I compiled all of the iterations of the deliverable in a .RAR directory for ease of access.  To extract the files within, download WinRAR or WinZip, right-click the directory, and choose extract.